Adds conversation timeout setting.
Introduces a conversation timeout setting to clear chat history after a specified duration, helping users manage costs and privacy.

The timeout can be configured in the General Settings tab and defaults to 5 minutes. A setting of 'Never' will disable the timeout. A warning message is displayed when timeout is set to never.

The SpeechService is updated to clear conversation history if the timeout has expired.